Tweak the html block's editor styles
