+++ packages/editor/src/index.js
@@ -16,4 +16,4 @@ import './hooks';
 
 export * from './components';
 export * from './utils';
-export * from './editor-styles';
+export { default as transformStyles } from './editor-styles';
